refactor: Cosmos indexers (#4394)
### Description

Refactoring Cosmos related indexers:
* Separate dispatch and deliver indexers
* Re-use repeated code on log parsing

### Backward compatibility

Yes

### Testing

E2E Test for Cosmos

---------
