+++ rust/chains/hyperlane-cosmos/src/merkle_tree_hook.rs
@@ -11,6 +11,7 @@ use once_cell::sync::Lazy;
 use tendermint::abci::EventAttribute;
 use tracing::instrument;
 
+use crate::utils::parse_logs_in_range;
 use crate::{
     grpc::WasmProvider,
     payloads::{general, merkle_tree_hook},
@@ -283,22 +284,12 @@ impl Indexer<MerkleTreeInsertion> for CosmosMerkleTreeHookIndexer {
         &self,
         range: RangeInclusive<u32>,
     ) -> ChainResult<Vec<(Indexed<MerkleTreeInsertion>, LogMeta)>> {
-        let logs_futures: Vec<_> = range
-            .map(|block_number| {
-                let self_clone = self.clone();
-                tokio::spawn(async move {
-                    let logs = self_clone
-                        .indexer
-                        .get_logs_in_block(
-                            block_number,
-                            Self::merkle_tree_insertion_parser,
-                            "MerkleTreeInsertionCursor",
-                        )
-                        .await;
-                    (logs, block_number)
-                })
-            })
-            .collect();
+        let logs_futures = parse_logs_in_range(
+            range,
+            self.indexer.clone(),
+            Self::merkle_tree_insertion_parser,
+            "MerkleTreeInsertionCursor",
+        );
 
         execute_and_parse_log_futures(logs_futures).await
     }
