+++ rust/chains/hyperlane-cosmos/src/utils.rs
@@ -1,13 +1,20 @@
+use std::fmt::Debug;
 use std::num::NonZeroU64;
+use std::ops::RangeInclusive;
 
 use base64::{engine::general_purpose::STANDARD as BASE64, Engine};
 use futures::future;
-use hyperlane_core::{ChainCommunicationError, ChainResult, Indexed, LogMeta};
 use once_cell::sync::Lazy;
+use tendermint::abci::EventAttribute;
 use tokio::task::JoinHandle;
 use tracing::warn;
 
+use hyperlane_core::{ChainCommunicationError, ChainResult, Indexed, LogMeta};
+
 use crate::grpc::{WasmGrpcProvider, WasmProvider};
+use crate::rpc::{CosmosWasmIndexer, ParsedEvent, WasmIndexer};
+
+type FutureChainResults<T> = Vec<JoinHandle<(ChainResult<Vec<(T, LogMeta)>>, u32)>>;
 
 /// The event attribute key for the contract address.
 pub(crate) const CONTRACT_ADDRESS_ATTRIBUTE_KEY: &str = "_contract_address";
@@ -32,6 +39,23 @@ pub(crate) async fn get_block_height_for_lag(
     };
 
     Ok(block_height)
+}
+
+pub(crate) fn parse_logs_in_range<T: PartialEq + Send + Sync + Debug + 'static>(
+    range: RangeInclusive<u32>,
+    indexer: Box<CosmosWasmIndexer>,
+    parser: for<'a> fn(&'a Vec<EventAttribute>) -> ChainResult<ParsedEvent<T>>,
+    label: &'static str,
+) -> FutureChainResults<T> {
+    range
+        .map(|block_number| {
+            let indexer = indexer.clone();
+            tokio::spawn(async move {
+                let logs = indexer.get_logs_in_block(block_number, parser, label).await;
+                (logs, block_number)
+            })
+        })
+        .collect()
 }
 
 #[allow(clippy::type_complexity)]
